requirements.txt
boto3==1.34.105
botocore==1.34.105
Flask==3.0.3
openai==1.30.2
Pillow==10.3.0
pillow_heif==0.16.0
Requests==2.32.2
urllib3~=2.2.1
elevenlabs~=1.2.2
python-dotenv~=1.0.1
pydub~=0.25.1
PyYAML~=6.0.1